% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_2by2.R
\name{calc_2by2}
\alias{calc_2by2}
\title{Calculate statistoics for 2 x 2 table}
\usage{
calc_2by2(outcome, exposure, col_names, conf_level)
}
\arguments{
\item{outcome}{Logical vector for outcome.}

\item{exposure}{Logical vector for exposure}

\item{col_names}{Character vector, length 2, to name outcome and exposure}

\item{conf_level}{Probability for confidence interval calculations}
}
\value{
Returns an epi_2by2 object.
}
\description{
Given two logical vectors calculate relevant 2 x 2 statistocs
}
\examples{
head(mtcars)
epi_2by2(mtcars, hp>100, am==1)
}
