% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_2by2.R
\name{epi_2by2}
\alias{epi_2by2}
\title{Two by two table analysis}
\usage{
epi_2by2(df, outcome, exposure)
}
\arguments{
\item{df}{A dataframe}

\item{outcome}{The unquoted column name for outcome.}

\item{exposure}{The unquoted column name for exposure}
}
\value{
Returns an epi_2by2 object.
}
\description{
A technology test rather than a final product
}
\details{
Introduce S3 class epi and explore printing and plotting methods
}
\examples{
epi_2by2(mtcars, gear, carb)
}
