% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_binom.R
\name{epi_binom}
\alias{epi_binom}
\title{Binomial estimates from dataframe column}
\usage{
epi_binom(x, y, conf_level = 0.95, methods = "all", digits, ...)
}
\arguments{
\item{x}{A dataframe or number of sucesses}

\item{y}{An expression evaluating to TRUE/FALSE or number of trials}

\item{conf_level}{Probability for confidence interval calculations}

\item{methods}{Confidence interval method see ?binom::binom.confint}

\item{digits}{Rounding digits}

\item{...}{Additional arguments for binom::binom.confint "bayes" method}
}
\value{
Returns an epi_binom object - this includes a variable for conf_level.

\if{html}{
\out{
 <img src = "https://media.giphy.com/media/IB9foBA4PVkKA/giphy.gif">
}}
}
\description{
A wrapper for binom::binom.confint  called with a data.frame and
unquoted columns or expressions.
}
\details{
Given a column of a dataframe (and 'positive' level) generate a
binomial proportion estimate and confidence intervals
}
\examples{
head(mtcars)
epi_binom(mtcars, cyl==4, conf_level = 0.95, methods = "all")
epi_binom(7, 10, conf_level = 0.95, methods = "all")
}
