% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_inc.R
\name{epi_inc}
\alias{epi_inc}
\title{Incidence rate estimates from dataframe column}
\usage{
epi_inc(df, event, time, conf_level = 0.95, method = "exact", digits)
}
\arguments{
\item{df}{A dataframe}

\item{event}{The unquoted event count column}

\item{time}{The unquoted tiem at risk column}

\item{conf_level}{Probability for confidence interval calculations}

\item{method}{The method to send to epiR::epi.conf}

\item{digits}{Rounding digits}
}
\value{
Returns an epi_inc object - this includes a variable for conf_level.
}
\description{
A wrapper for epiR::epi.conf  called with a data.frame and
unquoted columns or expressions.
}
\details{
Given a column of a incidence and time at risk generate an
estimate and confidence intervals
}
\examples{
head(mtcars)
epi_inc(mtcars, cyl, hp, conf_level = 0.95)
}
