% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_predval.R
\name{epi_predval}
\alias{epi_predval}
\title{Generate ppv and npv from test evaluation and prevalence range}
\usage{
epi_predval(X = NULL, Y = NULL, se = NULL, sp = NULL,
  prevalences = seq(0, 1, 0.05))
}
\arguments{
\item{X}{An epiR::epi.tests object or test sensitivity}

\item{Y}{Test specificity if not explicitly assigned to sp@param}

\item{se}{optionally test sensitivity}

\item{sp}{optionally test specificituy}

\item{prevalences}{vector of prevalences to calculate npv and ppv at}
}
\value{
Returns an epi_predval object
}
\description{
Generate ppv and npv from test evaluation and prevalence range
}
\examples{
head(mtcars)
res <- epi_tests(mtcars, mpg < 25, cyl > 4, conf.level = 0.95)
epi_predval(res)
}
