% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{epi_table}
\alias{epi_table}
\title{Utility function to convert df and columns to matrix for epiR functions}
\usage{
epi_table(df, outcome, exposure, outcome_labels = c("diseased", "healthy"),
  exposure_labels = c("exposed", "unexposed"))
}
\arguments{
\item{df}{dataframe to source 2 x 2 table}

\item{outcome}{The outcome column  (unquoted) or expression}

\item{exposure}{The exposure column (unquoted) or expresssion}

\item{outcome_labels}{A character vector, length 2, of outcome (positive, negative)}

\item{exposure_labels}{A character vectoe, length 2, of exposure (positive, negative)}
}
\description{
Utility function to convert df and columns to matrix for epiR functions
}
\examples{
head(mtcars)
tab <- epi_table(mtcars, am == 1, cyl == 4)
epiR::epi.2by2(tab, method = "cross.sectional")
epi_2by2(tab, method = "cross.sectional")
epi_2by2(mtcars, am == 1, cyl == 4, method = "cross.sectional")
}
