% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tests.R
\name{plot.epi.tests}
\alias{plot.epi.tests}
\title{Plot the output from epiR::epi.tests}
\usage{
\method{plot}{epi.tests}(x, ...)
}
\arguments{
\item{x}{An epi.tests object}

\item{...}{other arguments - ignored}
}
\value{
Returns a ggplot
}
\description{
Plot the output from epiR::epi.tests
}
\examples{
head(mtcars)
res <- epi_tests(mtcars, am == 1, cyl == 4, conf.level = 0.95)
plot(res)
}
