% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_predval.R
\name{plot.epi_predval}
\alias{plot.epi_predval}
\title{Default plotting of a epi_predval object}
\usage{
\method{plot}{epi_predval}(x, y = NULL, ...)
}
\arguments{
\item{x}{An epi_predval object to plot}

\item{y}{ignored}

\item{...}{ignored}
}
\value{
ggplot object
}
\description{
Default plotting of a epi_predval object
}
\examples{
pv <- epi_predval(se = 0.90, sp = 0.99)
plot(pv)
}
