% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot}
\alias{plot}
\title{Two by two table plot}
\usage{
plot(object)
}
\arguments{
\item{object}{An epi_2by2 object}

\item{colours}{A palette of four colours}
}
\description{
A technology test rather than a final product
}
\details{
Introduce S3 class epi and explore printing and plotting methods

This needs A LOT of work
}
\examples{
head(mtcars)
tab <- epi_2by2(mtcars, hp>100, am==1)
plot(tab)
}
