% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot.epi_2by2.R
\name{qplot}
\alias{qplot}
\title{Two by two table plot}
\usage{
qplot(object)
}
\arguments{
\item{object}{An epi_2by2 object}
}
\description{
A technology test rather than a final product
}
\details{
Introduce S3 class epi and explore printing and plotting methods
}
\examples{
tab <- epi_2by2(mtcars, gear, carb)
qplot(tab)
}
