% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.epi_2by2}
\alias{summary.epi_2by2}
\title{Two by two table summary}
\usage{
\method{summary}{epi_2by2}(object)
}
\arguments{
\item{object}{An epi_2by2 object}
}
\value{
Returns a sumamry of aepi_2by2 object.
}
\description{
A technology test rather than a final product
}
\details{
Summary method for epi_2by2 class

This needs A LOT of work
}
\examples{
head(mtcars)
epi_2by2(mtcars, hp>100, am==1)
#' summary(ep)
}
