% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TF_to_posneg}
\alias{TF_to_posneg}
\title{Utility function to convert TRUE/FALSE to positive/negative}
\usage{
TF_to_posneg(x)
}
\arguments{
\item{x}{Input vector}
}
\description{
Utility function to convert TRUE/FALSE to positive/negative
}
