% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_2by2.R
\name{tidy.epi.2by2}
\alias{tidy.epi.2by2}
\title{Tidy epi.2by2 output to give data.frame of estimates and CI}
\usage{
\method{tidy}{epi.2by2}(x, ...)
}
\arguments{
\item{x}{An epi.2by2 object}

\item{...}{other arguments (ignored)}
}
\description{
Tidy epi.2by2 output to give data.frame of estimates and CI
}
