% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tests.R
\name{tidy.epi.tests}
\alias{tidy.epi.tests}
\title{Tidy the output from epiR::epi.tests into a tibble}
\usage{
\method{tidy}{epi.tests}(x, ...)
}
\arguments{
\item{x}{An epi.tests object}

\item{...}{other arguments - ignored}
}
\value{
Returns a tibble
}
\description{
Tidy the output from epiR::epi.tests into a tibble
}
\examples{
head(mtcars)
res <- epi_tests(mtcars, mpg < 25, cyl > 4, conf.level = 0.95)
tidy(res)
}
