% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tests.R
\name{twobytwo}
\alias{twobytwo}
\title{Make an epi_twobytwo object for epi_tests}
\usage{
twobytwo(a, b, c, d)
}
\arguments{
\item{a}{The TRUE POSITIVE count}

\item{b}{The FALSE POSITIVE count}

\item{c}{The FALSE NEGATIVE count}

\item{d}{The TRUE NEGATIVE count}
}
\value{
Returns an epi.tests
}
\description{
A helper function for epi_tests that allows entry of summary data
}
\examples{
epi_tests(twobytwo(12, 14, 1, 5))
}
