% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_2by2.R
\name{epi_2by2}
\alias{epi_2by2}
\title{Two by two table analysis
#'
Given two columns of a dataframe (and 'positive' levels) generate a 2 x 2
table and appropriate summary statsitics}
\usage{
epi_2by2(df, outcome, exposure, outcome_positive = TRUE,
  exposure_positive = TRUE, method = "cohort.count", conf_level = 0.95,
  homogeneity = "breslow.day", outcome_format = "as.columns")
}
\arguments{
\item{df}{A dataframe}

\item{outcome}{The unquoted column name for outcome.}

\item{exposure}{The unquoted column name for exposure}

\item{outcome_positive}{The "level" for a positive / of interest outcome}

\item{exposure_positive}{The "level" for a positive / of interest exposure}

\item{method}{The study type "cross", "cohort" or "case-control"}

\item{conf_level}{Probability for confidence interval calculations}

\item{homogeneity}{TODO}

\item{outcome_format}{TODO}
}
\value{
Returns an epi_2by2 object.
}
\description{
Two by two table analysis
#'
Given two columns of a dataframe (and 'positive' levels) generate a 2 x 2
table and appropriate summary statsitics
}
\examples{
head(mtcars)
epi_2by2(mtcars, hp>100, am==1)
}
