% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_2by2.R
\name{epi_2by2}
\alias{epi_2by2}
\title{Wrap data from a dataframe and send to epiR::epi.2by2 (or pass through table)}
\usage{
epi_2by2(x, outcome = NULL, exposure = NULL, ...)
}
\arguments{
\item{x}{A dataframe or numeric vector length 4}

\item{outcome}{The unquoted column name for test results}

\item{exposure}{The unquoted column name for exposure column}

\item{...}{Other parameters passed on to epiR::epi.2by2}
}
\value{
Returns an epi.2by2 object
}
\description{
A wrapper for epirR::epi.2by2 normally called with a data.frame and
unquoted columns or expressions. Can also be called with a numeric vector length 4 to allow direct entry of 2 x 2 table data
}
\examples{
head(mtcars)
epi_2by2(mtcars, am == 1, cyl == 4, conf.level = 0.95)
epi_2by2(c(12, 34, 1, 6), conf.level = 0.95)
}
