% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_binom.R
\name{epi_binom}
\alias{epi_binom}
\title{Binomial estimates from dataframe column}
\usage{
epi_binom(df, var, var_positive = TRUE, conf_level = 0.95,
  methods = "all", digits, ...)
}
\arguments{
\item{df}{A dataframe}

\item{var}{The unquoted column name for variable}

\item{var_positive}{The "level" for a positive / of interest outcome}

\item{conf_level}{Probability for confidence interval calculations}

\item{methods}{Confidence interval method see ?binom::binom.confint}

\item{digits}{Rounding digits}

\item{...}{Additional arguments for binom::binom.confint "bayes" method}
}
\value{
Returns an epi_binom object - this includes a variable for conf_level.
}
\description{
A wrapper for binom::binom.confint  called with a data.frame and
unquoted columns or expressions.
}
\details{
Given a column of a dataframe (and 'positive' level) generate a
binomial proportion estimate and confidence intervals
}
\examples{
head(mtcars)
epi_binom(mtcars, cyl, 4, conf_level = 0.95, methods = "all")
}
