% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_inc.R
\name{epi_inc}
\alias{epi_inc}
\title{Estimate incidence rate}
\usage{
epi_inc(x, y, z = NULL, conf_level = 0.95, methods = "exact", ...)
}
\arguments{
\item{x}{A dataframe or number of cases}

\item{y}{See details}

\item{z}{See details}

\item{conf_level}{confidence level for estimates}

\item{methods}{estimation methods for CI to be passes to epiR::epi_conf}

\item{...}{Other parameters passed on to epiR::epi.conf}
}
\value{
Returns a tibble
}
\description{
Wrap data from a dataframe and send to epiR::epi.conf (or pass through table).
To calculate an incidence rate using cases and time at risk
}
\details{
If used with a dataframe use epi_inc(dataframe, cases-expression, time_at_risk-column, ...)

If used with summary data use epi_inc(number-of-cases, time-at-risk, ...)
}
\examples{
head(mtcars)
# using a dataframe...
epi_inc(mtcars, cyl == 4, disp, conf_level = 0.95)
# using summary data
epi_inc(70, 1010, conf_level = 0.95)
}
