% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tests.R
\name{epi_tests}
\alias{epi_tests}
\title{Wrap data from a dataframe and send to epiR::epi.tests}
\usage{
epi_tests(x, test = NULL, goldstandard = NULL, conf_level = 0.95)
}
\arguments{
\item{x}{A dataframe or epi_two_by_two object}

\item{test}{The unquoted column name for test results}

\item{goldstandard}{The unquoted column name for gold standard test results}

\item{conf_level}{Probability for confidence interval calculations}
}
\value{
Returns an epi.tests
}
\description{
A wrapper for epirR::epi.tests normally called with a data.frame and
unquoted columns or expressions. Can also be called with an epi_twobytwo object (generated with twobytwo helper function) to allow direct entry of a table of results.
}
\examples{
head(mtcars)
epi_tests(mtcars, mpg < 25, cyl > 4, conf_level = 0.95)
}
