% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_tests.R
\name{epi_tests}
\alias{epi_tests}
\title{Wrap data from a dataframe and send to epiR::epi.tests}
\usage{
epi_tests(x, test = NULL, goldstandard = NULL, ...)
}
\arguments{
\item{x}{A dataframe or numeric vector length 4}

\item{test}{The unquoted column name for test results}

\item{goldstandard}{The unquoted column name for gold standard test results}

\item{...}{Other parameters passed on to epiR::epi.tests}
}
\value{
Returns an epi.tests
}
\description{
A wrapper for epirR::epi.tests normally called with a data.frame and
unquoted columns or expressions. Can also be called with a numeric vector length 4 to allow direct entry of 2 x 2 table data
}
\examples{
head(mtcars)
epi_tests(mtcars, mpg < 25, cyl > 4, conf.level = 0.95)
}
