% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.epi_2by2}
\alias{print.epi_2by2}
\title{Print method for epi_2by2 class}
\usage{
\method{print}{epi_2by2}(object)
}
\arguments{
\item{object}{An epi_2by2 object}
}
\description{
A technology test rather than a final product
}
\examples{
library(epi_2by2)
head(mtcars)
tab <- epi_2by2(mtcars, hp>100, am==1)
print(tab)
}
